Move controller for admin delete variables to ProcessInstanceVariableAdminController
