+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-rest-api/src/main/java/org/activiti/cloud/services/rest/api/ProcessInstanceVariableAdminController.java
@@ -18,10 +18,12 @@ package org.activiti.cloud.services.rest.api;
 
 import java.util.List;
 
+import org.activiti.api.process.model.payloads.RemoveProcessVariablesPayload;
 import org.activiti.api.process.model.payloads.SetProcessVariablesPayload;
 import org.springframework.hateoas.MediaTypes;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -37,6 +39,15 @@ public interface ProcessInstanceVariableAdminController {
      * @return list of error messages, empty list if no error message
      */
     @PutMapping
-    ResponseEntity<List<String>> setVariables(@PathVariable String processInstanceId,
+    ResponseEntity<List<String>> updateVariables(@PathVariable String processInstanceId,
                                               @RequestBody SetProcessVariablesPayload setProcessVariablesPayload);
+    
+    /**
+     * @param processInstanceId          id of process instance
+     * @param removeProcessVariablesPayload process variables payload
+     */
+    @DeleteMapping
+    ResponseEntity<Void> removeVariables(@PathVariable String processInstanceId,
+                                         @RequestBody RemoveProcessVariablesPayload removeProcessVariablesPayload);
+
 }
