+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-rest-impl/src/test/java/org/activiti/cloud/services/rest/controllers/ProcessInstanceVariableAdminControllerImplIT.java
@@ -20,12 +20,16 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.put;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.activiti.api.process.model.builders.ProcessPayloadBuilder;
 import org.activiti.api.process.runtime.ProcessRuntime;
 import org.activiti.api.runtime.conf.impl.CommonModelAutoConfiguration;
@@ -59,8 +63,6 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.ResultActions;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 @RunWith(SpringRunner.class)
 @WebMvcTest(ProcessInstanceVariableAdminControllerImpl.class)
@@ -252,5 +254,19 @@ public class ProcessInstanceVariableAdminControllerImplIT {
         assertThat(actualResponseBody).contains(expectedTypeErrorMessage);
         assertThat(actualResponseBody).contains(expectedNameErrorMessage1);
         assertThat(actualResponseBody).contains(expectedNameErrorMessage2);
+    }
+    
+    @Test
+    public void deleteVariables() throws Exception {
+        this.mockMvc.perform(delete("/admin/v1/process-instances/{processInstanceId}/variables",
+                                    "1")
+                                    .accept(MediaTypes.HAL_JSON_VALUE)
+                                    .contentType(MediaType.APPLICATION_JSON)
+                                    .content(mapper.writeValueAsString(ProcessPayloadBuilder.removeVariables().withVariableNames(Arrays.asList("varName1",
+                                                                                                                                               "varName2"))
+                                    .build())))
+                .andDo(print())
+                .andExpect(status().isOk());
+        verify(processAdminRuntime).removeVariables(any());
     }
 }
